Apply fixed Coderabbitai review.
