+++ src/Form/InputCheckbox.php
@@ -11,15 +11,12 @@ use UIAwesome\Html\Attribute\HasValue;
 use UIAwesome\Html\Attribute\Values\Type;
 use UIAwesome\Html\Core\Element\BaseInput;
 use UIAwesome\Html\Core\Html;
-use UIAwesome\Html\Form\Mixin\{CanBeEnclosedByLabel, HasLabel};
-use UIAwesome\Html\Helper\Enum;
+use UIAwesome\Html\Form\Mixin\{CanBeEnclosedByLabel, HasCheckedState, HasLabel};
 use UIAwesome\Html\Interop\{VoidInterface, Voids};
 use UIAwesome\Html\Phrasing\Label;
 use UnitEnum;
 
-use function is_array;
-use function is_bool;
-use function is_scalar;
+use function array_key_exists;
 
 /**
  * Represents the HTML `<input type="checkbox">` element.
@@ -45,17 +42,11 @@ final class InputCheckbox extends BaseInput
 {
     use CanBeAutofocus;
     use CanBeEnclosedByLabel;
+    use HasCheckedState;
     use HasLabel;
     use HasRequired;
     use HasTabindex;
     use HasValue;
-
-    /**
-     * Determines the checked state of the checkbox.
-     *
-     * @phpstan-var mixed[]|bool|float|int|string|Stringable|UnitEnum|null $checked
-     */
-    private array|bool|float|int|string|Stringable|UnitEnum|null $checked = null;
 
     /**
      * Value to be submitted when the checkbox is not checked.
@@ -65,37 +56,6 @@ final class InputCheckbox extends BaseInput
      * This ensures that a value is always submitted for the checkbox, even when it is unchecked.
      */
     private bool|float|int|string|Stringable|UnitEnum|null $uncheckedValue = null;
-
-    /**
-     * Sets the `checked` attribute.
-     *
-     * Usage example:
-     * ```php
-     * $element->checked(true);
-     * $element->checked(false);
-     * $element->checked('active')->value('active'); // checked
-     * $element->checked('inactive')->value('active'); // unchecked
-     * ```
-     *
-     * @param array|bool|float|int|string|Stringable|UnitEnum|null $value Checked state.
-     *
-     * - `array`: Checkbox is checked if the value is in the array.
-     * - `false`: Checkbox is unchecked.
-     * - `true`: Checkbox is checked.
-     * - `float|int|string|Stringable|UnitEnum`: Checkbox is checked if the value matches the `value` attribute.
-     * - `null`: Attribute is removed.
-     *
-     * @return static New instance with the updated `checked` attribute.
-     *
-     * @phpstan-param mixed[]|bool|float|int|string|Stringable|UnitEnum|null $value
-     */
-    public function checked(array|bool|float|int|string|Stringable|UnitEnum|null $value): static
-    {
-        $new = clone $this;
-        $new->checked = $value;
-
-        return $new;
-    }
 
     /**
      * Returns the array of HTML attributes for the element.
@@ -211,59 +171,5 @@ final class InputCheckbox extends BaseInput
                 '{unchecked}' => $unchecked,
             ],
         );
-    }
-
-    /**
-     * Builds the attributes for the `<input>` element.
-     *
-     * This method normalizes the `value` attribute and determines the `checked` state based on the current value and
-     * the configured checked options.
-     *
-     * @param array $attributes Initial attributes array.
-     *
-     * @return array Updated attributes array with the `checked` attribute if applicable.
-     *
-     * @phpstan-param mixed[] $attributes
-     * @phpstan-return mixed[]
-     */
-    private function buildAttributes(array $attributes): array
-    {
-        if (isset($attributes['value']) && is_bool($attributes['value'])) {
-            $attributes['value'] = $attributes['value'] ? 1 : 0;
-        }
-
-        $checked = $this->checked;
-
-        $normalizedChecked = is_array($checked)
-            ? Enum::normalizeArray($checked)
-            : Enum::normalizeValue($checked);
-
-        if ($normalizedChecked === false || $normalizedChecked === null) {
-            return $attributes;
-        }
-
-        $value = $attributes['value'] ?? null;
-
-        if ($normalizedChecked === true) {
-            $attributes['checked'] = true;
-
-            return $attributes;
-        }
-
-        $valueStr = is_scalar($value) ? (string) $value : '';
-
-        if (is_array($normalizedChecked) === false) {
-            $attributes['checked'] = $valueStr === "{$normalizedChecked}";
-
-            return $attributes;
-        }
-
-        foreach ($normalizedChecked as $item) {
-            if (is_scalar($item) && "{$item}" === $valueStr) {
-                $attributes['checked'] = true;
-            }
-        }
-
-        return $attributes;
     }
 }
